skills/session-context/pyproject.toml
[project]
name = "session-context"
version = "0.1.0"
description = "Gathers session context: git history, recent plans, pending work"
requires-python = ">=3.11"
dependencies = [
    "gitpython>=3.1.40",
    "pyyaml>=6.0",
    "structlog>=24.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
